Move usecases folder
