+++ internal/api/api.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 
 	"github.com/rabbytesoftware/quiver/internal/api/middleware"
-	"github.com/rabbytesoftware/quiver/internal/api/v1/usecases"
+	"github.com/rabbytesoftware/quiver/internal/usecases"
 	"github.com/rabbytesoftware/quiver/internal/core/config"
 	"github.com/rabbytesoftware/quiver/internal/core/watcher"
