+++ internal/api/v1/routes/arrows/routes.go
@@ -3,7 +3,7 @@ package arrows
 import (
 	"github.com/gin-gonic/gin"
 	arrowhandler "github.com/rabbytesoftware/quiver/internal/api/v1/handlers/arrows"
-	usecase "github.com/rabbytesoftware/quiver/internal/api/v1/usecases/arrows"
+	usecase "github.com/rabbytesoftware/quiver/internal/usecases/arrows"
 )
 
 func SetupRoutes(router *gin.RouterGroup, usecases *usecase.ApiArrowUsecases) {
