+++ internal/api/v1/routes/quivers/routes.go
@@ -2,7 +2,7 @@ package quivers
 
 import (
 	"github.com/gin-gonic/gin"
-	usecase "github.com/rabbytesoftware/quiver/internal/api/v1/usecases/quivers"
+	usecase "github.com/rabbytesoftware/quiver/internal/usecases/quivers"
 )
 
 func SetupRoutes(
