+++ internal/api/v1/routes/system/routes.go
@@ -2,7 +2,7 @@ package system
 
 import (
 	"github.com/gin-gonic/gin"
-	usecase "github.com/rabbytesoftware/quiver/internal/api/v1/usecases/system"
+	usecase "github.com/rabbytesoftware/quiver/internal/usecases/system"
 )
 
 func SetupRoutes(router *gin.RouterGroup, usecases *usecase.ApiSystemUsescases) {
