+++ internal/internal.go
@@ -2,7 +2,7 @@ package internal
 
 import (
 	"github.com/rabbytesoftware/quiver/internal/api"
-	"github.com/rabbytesoftware/quiver/internal/api/v1/usecases"
+	"github.com/rabbytesoftware/quiver/internal/usecases"
 	"github.com/rabbytesoftware/quiver/internal/core"
 	"github.com/rabbytesoftware/quiver/internal/infrastructure"
 	"github.com/rabbytesoftware/quiver/internal/repositories"
