--- internal/usecases/quivers/usecase.go
@@ -1,15 +1,0 @@
-package quivers
-
-import "github.com/rabbytesoftware/quiver/internal/repositories"
-
-type QuiversUsecase struct {
-	repositories *repositories.Repositories
-}
-
-func NewQuiversUsecase(
-	repositories *repositories.Repositories,
-) *QuiversUsecase {
-	return &QuiversUsecase{
-		repositories: repositories,
-	}
-}
