+++ internal/usecases/usecases.go
@@ -1,24 +1,22 @@
 package usecases
 
 import (
+	"github.com/rabbytesoftware/quiver/internal/api/v1/usecases/arrows"
+	"github.com/rabbytesoftware/quiver/internal/api/v1/usecases/quivers"
+	"github.com/rabbytesoftware/quiver/internal/api/v1/usecases/system"
 	"github.com/rabbytesoftware/quiver/internal/repositories"
-	"github.com/rabbytesoftware/quiver/internal/usecases/arrows"
-	"github.com/rabbytesoftware/quiver/internal/usecases/quivers"
-	"github.com/rabbytesoftware/quiver/internal/usecases/system"
 )
 
-type Usecases struct {
-	Arrows  *arrows.ArrowsUsecase
-	Quivers *quivers.QuiversUsecase
-	System  *system.SystemUsecase
+type ApiUsescases struct {
+	Arrows  *arrows.ApiArrowUsecases
+	Quivers *quivers.ApiQuiverUsescases
+	System  *system.ApiSystemUsescases
 }
 
-func NewUsecases(
-	repositories *repositories.Repositories,
-) *Usecases {
-	return &Usecases{
-		Arrows:  arrows.NewArrowsUsecase(repositories),
-		Quivers: quivers.NewQuiversUsecase(repositories),
-		System:  system.NewSystemUsecase(repositories),
+func NewApiUsecases(repos *repositories.Repositories) *ApiUsescases {
+	return &ApiUsescases{
+		Arrows:  arrows.NewApiArrowsUsecases(repos),
+		Quivers: quivers.NewApiQuiversUsecases(repos),
+		System:  system.NewApiSystemUsecases(repos),
 	}
 }
